Last changes

+++ src/cars/brand/model/GasCar.java
@@ -1,13 +1,11 @@
 package cars.brand.model;
 
-import cars.brand.model.Vehicle;
-
 public class GasCar extends Vehicle {
 	
-	protected double engineDisplacement;
+	protected float engineDisplacement;
 	protected double price;
 
-	public GasCar(String type, String brand, String modelName, double engineDisplacement, short enginePower, double price) {
+	public GasCar(String type, String brand, String modelName, float engineDisplacement, short enginePower, double price) {
 		super(type, brand, modelName, enginePower);
 		this.engineDisplacement = engineDisplacement;
 		this.price = price;
@@ -23,20 +21,18 @@ public class GasCar extends Vehicle {
 		return super.equals(vehicle);
 	}
 
-	public double getEngineDisplacement() {
+	public float getEngineDisplacement() {
 		return engineDisplacement;
 	}
 
-	public void setEngineDisplacement(double engineDisplacement) {
+	public void setEngineDisplacement(float engineDisplacement) {
 		this.engineDisplacement = engineDisplacement;
 	}
-
 	
 	public double getPrice() {
 		return price;
 	}
 
-	
 	public void setPrice(double price) {
 		this.price = price;
 	}
